# docs/openapi.yaml
openapi: 3.0.3
info:
  title: F4ALL Sports Talent Assessment API
  description: >
    API contract for the F4ALL AI-powered sports talent assessment platform.
    The mobile application receives a provisional on-device score, while the
    backend performs server-side verification and stores the final result.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/auth/request-otp:
    post:
      tags:
        - Authentication
      summary: Request OTP
      description: Send an OTP to the athlete's registered phone number.

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestOtpRequest'

      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

        '400':
          description: Invalid phone number
  /api/auth/verify-otp:
    post:
      tags:
        - Authentication
      summary: Verify OTP
      description: Verify the OTP and authenticate the athlete.

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpRequest'

      responses:
        '200':
          description: OTP verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

        '400':
          description: Invalid or expired OTP


  # ============================================================
  # TEST SUBMISSION
  # ============================================================
...
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Exchange a refresh token for a new token pair
      description: >
        Rotates the refresh token. Presenting an already-rotated token revokes
        every session for that subject, since two parties must be holding it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: {type: string}
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TokenPair"}
        "401": {description: Unknown, expired, revoked or reused refresh token}
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke a refresh token, or every device's
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: {type: string, nullable: true}
                all_devices: {type: boolean, default: false}
      responses:
        "200": {description: Signed out}
...
  /api/dashboard/auth/refresh:
    post:
      tags: [Dashboard]
      summary: Rotate an official refresh token
      description: Athlete refresh tokens are refused.
      responses:
        "200": {description: New token pair}
        "401": {description: Invalid token or inactive account}
...
components:
...
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT


  # ------------------------------------------------------------
  # REQUEST SCHEMAS
  # ------------------------------------------------------------
  schemas:
    RequestOtpRequest:
      type: object
      required:
        - phone
      properties:
        phone:
          type: string
          description: Athlete phone number
          example: '+919876543210'
    VerifyOtpRequest:
      type: object
      required:
        - phone
        - otp
      properties:
        phone:
          type: string
          description: Athlete phone number
          example: '+919876543210'

        otp:
          type: string
          description: One-time password
          example: '123456'
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: OTP sent successfully
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: jwt-token

        athlete_id:
          type: string
          format: uuid
          description: Authenticated athlete ID
          example: '550e8400-e29b-41d4-a716-446655440000'
...
    TokenPair:
      type: object
      properties:
        access_token: {type: string}
        refresh_token: {type: string}
        token_type: {type: string, example: bearer}
        expires_in: {type: integer, description: Access token lifetime in seconds}
        athlete_id: {type: string, format: uuid, nullable: true}
        registered: {type: boolean}
...